+++ server/src/main/java/nl/tudelft/oopp/group54/controllers/lectures/LectureService.java
@@ -6,7 +6,7 @@ import java.util.Map;
 public interface LectureService {
     Map<String, Object> createNewLecture(Date startTime, String lectureName);
 
-    Map<String, Object> joinOngoingLecture(Long lectureId, Long joinId, String userName);
+    Map<String, Object> joinOngoingLecture(Integer lectureId, Long joinId, String userName);
 
-    Map<String, Object> getLectureMetadata(Long lectureId);
+    Map<String, Object> getLectureMetadata(Integer lectureId);
 }
